begin moving reducers into logically correct place
