+++ packages/space-opera/src/components/camera_settings/camera_settings_test.ts
@@ -18,11 +18,12 @@
 
 import './camera_settings.js';
 
+import {dispatchAutoRotate, dispatchCameraTarget, dispatchInitialOrbit} from '../../redux/reducers/camera_settings.js';
 import {dispatchCurrentCameraState} from '../../redux/space_opera_base.js';
 import {reduxStore} from '../../redux/space_opera_base.js';
 import {Vector3D} from '../../redux/state_types.js';
 
-import {CameraSettings, CameraTargetInput, dispatchAutoRotate, dispatchCameraTarget, dispatchInitialOrbit} from './camera_settings.js';
+import {CameraSettings, CameraTargetInput} from './camera_settings.js';
 
 describe('camera constraints test', () => {
   let cameraSettings: CameraSettings;
