+++ packages/space-opera/src/components/poster_controls/poster_controls.ts
@@ -25,8 +25,8 @@ import {ModelViewerElement} from '@google/model-viewer/lib/model-viewer';
 import {customElement, html, internalProperty} from 'lit-element';
 
 import {dispatchSetPoster} from '../../redux/poster_dispatchers.js';
+import {dispatchSaveCameraOrbit} from '../../redux/reducers/camera_settings.js';
 import {State} from '../../redux/space_opera_base.js';
-import {dispatchSaveCameraOrbit} from '../camera_settings/camera_settings.js';
 import {ConnectedLitElement} from '../connected_lit_element/connected_lit_element.js';
 
 import {styles} from './poster_controls.css';
